- Updated Spring data Commons and Spring data build to the current snapshot - Moved the query annotations to neo4j-spring as extra spring annotation required on @Query to make it work. - Refactored locations of current classes to be more consistent with current SDN as well as other SD projects like JPA.
