+++ neo4j-spring/src/main/java/org/springframework/data/neo4j/repository/config/Neo4jRepositoryConfigurationExtension.java
@@ -1,6 +1,6 @@
-package org.springframework.data.neo4j.support;
+package org.springframework.data.neo4j.repository.config;
 
-import org.springframework.data.neo4j.repository.GraphRepositoryFactoryBean;
+import org.springframework.data.neo4j.repository.support.GraphRepositoryFactoryBean;
 import org.springframework.data.repository.config.RepositoryConfigurationExtensionSupport;
 
 public class Neo4jRepositoryConfigurationExtension extends RepositoryConfigurationExtensionSupport {
