+++ neo4j-spring/src/main/java/org/springframework/data/neo4j/repository/support/GraphRepositoryFactoryBean.java
@@ -1,7 +1,8 @@
-package org.springframework.data.neo4j.repository;
+package org.springframework.data.neo4j.repository.support;
 
 import org.neo4j.ogm.session.Session;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.neo4j.repository.support.GraphRepositoryFactory;
 import org.springframework.data.repository.Repository;
 import org.springframework.data.repository.core.support.RepositoryFactorySupport;
 import org.springframework.data.repository.core.support.TransactionalRepositoryFactoryBeanSupport;
