+++ neo4j-spring/src/test/java/org/springframework/data/neo4j/integration/movies/repo/UserRepository.java
@@ -1,6 +1,6 @@
 package org.springframework.data.neo4j.integration.movies.repo;
 
-import org.neo4j.ogm.annotation.Query;
+import org.springframework.data.neo4j.annotation.Query;
 import org.springframework.data.neo4j.integration.movies.domain.User;
 import org.springframework.data.neo4j.repository.GraphRepository;
 import org.springframework.stereotype.Repository;
