+++ neo4j-spring/src/test/java/org/springframework/data/neo4j/integration/web/context/WebPersistenceContext.java
@@ -4,7 +4,7 @@ import org.neo4j.ogm.session.Session;
 import org.neo4j.ogm.session.SessionFactory;
 import org.springframework.data.neo4j.integration.InProcessServer;
 import org.springframework.context.annotation.*;
-import org.springframework.data.neo4j.EnableNeo4jRepositories;
+import org.springframework.data.neo4j.repository.config.EnableNeo4jRepositories;
 import org.springframework.data.neo4j.config.Neo4jConfiguration;
 import org.springframework.data.neo4j.server.Neo4jServer;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
