+++ pom.xml
@@ -62,9 +62,15 @@
             </dependency>
 
             <dependency>
+                <groupId>org.springframework.data.build</groupId>
+                <artifactId>spring-data-parent</artifactId>
+                <version>1.6.0.BUILD-SNAPSHOT</version>
+            </dependency>
+
+            <dependency>
                 <groupId>org.springframework.data</groupId>
                 <artifactId>spring-data-commons</artifactId>
-                <version>1.6.1.RELEASE</version>
+                <version>1.10.0.BUILD-SNAPSHOT</version>
             </dependency>
 
             <dependency>
@@ -72,12 +78,6 @@
                 <artifactId>spring-test</artifactId>
                 <version>4.1.4.RELEASE</version>
                 <scope>test</scope>
-            </dependency>
-
-            <dependency>
-                <groupId>org.springframework.data</groupId>
-                <artifactId>spring-data-jpa</artifactId>
-                <version>1.4.1.RELEASE</version>
             </dependency>
 
             <dependency>
